refactored test structure of general tests
